+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-G5tJ8Ox_.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-Dx40hcf_.js";import{_ as a}from"./mermaid.core-DWicBCi9.js";import"./chunk-5VM5RSS4-DSuYcWIE.js";import"./chunk-XXDRQBXY-DAFBRngu.js";import"./chunk-POPQ4Y6H-Bt9y2-3J.js";import"./chunk-F27PBJKO-CwTtgNIH.js";import"./channel-BwjhNwjM.js";import"./index-B0Vs5m_n.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};